Fix check for body/html node type in _initDocument
The check in _initDocument that the node is of the right type will always fail,
since the node will never be both an instance of HTMLBodyElement and
HTMLHtmlElement. Fix this to only check against the type we expect based on
WHOLE_DOCUMENT cofig, and tidy code to be a little easier to read.