Update _isClobbered to check currently required properties
The _isClobbered fn detects whether any of the methods/attributes we need on
the element have been clobbered. The properties it checks for had not been
updated with the changes in how the sanitisation methods manipulated the
elements, meaning it was making some unnecessary checks, and not making some
required ones, which could result in an error being thrown with certain input.